Refactored examples into multiple services
